Overhaul of the datatransform testing module ...
Migrated both test files to leverage pytest. Moved the
data being stored in `keylookup_graphs` into fixtures within
the `conftest`. Removed the README as all the tests are now
marked with xfail with some indicators as to why they're failing
