--- tests/hub/datatransform/README
@@ -1,1 +1,0 @@
-non-functional tests...
